Rename processMessages.js to handleInput.js
and splits input handlers in processMessage.js to handlers/xxx.js
