+++ src/handlers/__tests__/utils.test.js
@@ -1,0 +1,10 @@
+describe('createPostbackAction()', () => {
+  it('should return postback message body');
+});
+
+describe('createFeedbackWords()', () => {
+  it('should create empty feedback words');
+  it('should create positive feedback words');
+  it('should create negative feedback words');
+});
+describe('createReferenceWords()', () => {});
